fix up tests
add a test wrapper to run the tests on Gauche. I think the test hook-run
fixup is correct since zero is the expected result of _not_ running the
hook.

Chibi also fails this test, but it's a bit harder to see the litte red
'x' (I missed it the first time, and thought something broke subtly on
Gauche)